Fold / Hopf Continuation

Fold / Hopf Continuation

For this to work, it is important to have an analytical expression for the jacobian. See the example Chan for more details.

The case of the Fold point

One a Fold point have been detected after a call to br, _ = continuation(...), it can be refined with the use of newton iterations. Several methods have been implemented namely Moore Spence and Minimally Augmented. A simplified interface is provided for the use of the later but the former one is fully functional.

Newton refinement

Let us say that ind_fold is the index in br.bifpoint of a Fold point. This guess can be refined by calling the simplified interface. More precisions are provided below for an advanced usage.

outfold, hist, flag = @time Cont.newtonFold((x,p) -> F(x, p),
							(x, p) -> Jac(x, p),
							br, ind_fold,
							opt_newton)

The case of the Hopf point

One a Hopf point have been detected after a call to br, _ = continuation(...), it can be refined with the use of newton iterations. Several method have been implemented but we focus on the Minimally Augmented one. A simplified interface is provided for the use of this method.

Newton refinement

Let us say that ind_hopf is the index in br.bifpoint of a Hopf point. This guess can be refined by calling the simplified interface. More precisions are provided below for an advanced usage.

outfold, hist, flag = @time Cont.newtonHopf((x,p) -> F(x, p),
							(x, p) -> Jac(x, p),
							br, ind_hopf,
							opt_newton)

Functions

PseudoArcLengthContinuation.newtonFold — Function.

This function turns an initial guess for a Fold point into a solution to the Fold problem based on a Minimally Augmented formulation. The arguments are as follows

(x, p)-> F(x, p) where p is the parameter associated to the Fold point

J = (x, p)-> d_xF(x, p) associated jacobian

Jt = (x, p) -> transpose(d_xF(x, p)) associated jacobian, it should be implenented otherwise

foldpointguess initial guess (x0, p0) for the Fold point. It should be a Vector

eigenvec guess for the 0 eigenvector

options::NewtonPar

source

Simplified call to refine an initial guess for a Fold point. More precisely, the call is as follows newtonFold(F, J, Jt, br::ContResult, index::Int64, options) where the parameters are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

source

PseudoArcLengthContinuation.newtonHopf — Function.

This function turns an initial guess for a Hopf point into a solution to the Hopf problem based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is the parameter associated to the Hopf point

J = (x, p) -> d_xF(x, p) associated jacobian

Jt = (x, p) -> transpose(d_xF(x, p)) associated jacobian

hopfpointguess initial guess (x0, p0) for the Hopf point. It should be a AbstractVector or a BorderedVector.

eigenvec guess for the iω eigenvector

eigenvec_ad guess for the -iω eigenvector

options::NewtonPar

source

Simplified call to refine an initial guess for a Hopf point. More precisely, the call is as follows newtonHopf(F, J, Jt, br::ContResult, index::Int64, options) where the parameters are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Eigenvectors`

This simplified call has been written when the eigenvectors are organised in a 2d Array evec where evec[:,2] is the second eigenvector in the list.

source

PseudoArcLengthContinuation.continuationFold — Function.

codim 2 continuation of Fold points. This function turns an initial guess for a Fold point into a curve of Fold points based on a Minimally Augmented formulation. The arguments are as follows

(x, p1, p2)-> F(x, p1, p2) where p is the parameter associated to the Fold point

J = (x, p1, p2)-> d_xF(x, p1, p2) associated jacobian

foldpointguess initial guess (x0, p10) for the Fold point. It should be a Vector

p2 parameter p2 for which foldpointguess is a good guess

eigenvec guess for the 0 eigenvector at p1_0

options::NewtonPar

source

Simplified call for continuation of Fold point. More precisely, the call is as follows continuationFold(F, J, Jt, br::ContResult, index::Int64, options) where the parameters are as for continuationFold except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

source

PseudoArcLengthContinuation.continuationHopf — Function.

codim 2 continuation of Hopf points. This function turns an initial guess for a Hopf point into a curve of Hopf points based on a Minimally Augmented formulation. The arguments are as follows

(x, p1, p2)-> F(x, p1, p2) where p is the parameter associated to the hopf point

J = (x, p1, p2)-> d_xF(x, p1, p2) associated jacobian

hopfpointguess initial guess (x0, p10) for the Hopf point. It should be a Vector or a BorderedVector

p2 parameter p2 for which hopfpointguess is a good guess

eigenvec guess for the iω eigenvector at p1_0

eigenvec_ad guess for the -iω eigenvector at p1_0

options::NewtonPar

source

Simplified call for continuation of Hopf point. More precisely, the call is as follows continuationHopf(F, J, Jt, br::ContResult, index::Int64, options) where the parameters are as for continuationHopf except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Eigenvectors`

This simplified call has been written when the eigenvectors are organised in a 2d Array evec where evec[:,2] is the second eigenvector in the list.

source