Library

Functions

PseudoArcLengthContinuation.newton — Function.

newton(F, J, x0, options, normN = norm)

This is the Newton Solver for F(x) = 0 with Jacobian J and initial guess x0. The function normN allows to specify a norm for the convergence criteria. It is important to set the linear solver options.linsolve properly depending on your problem. This solver is used to solve $J(x)u = -F(x)$ in the Newton step. You can for example use Default() which is the operator backslash which works well for Sparse / Dense matrices. Iterative solver (GMRES) are also implemented. You should implement your own for maximal efficiency. This is quite easy to do, have a look at src/LinearSolver.jl

Output:

solution

history of residuals

flag of convergence

number of iterations

source

PseudoArcLengthContinuation.newtonDeflated — Function.

newtonDeflated(Fhandle::Function, Jhandle, x0, options:: NewtonPar{T}, defOp::DeflationOperator{T, vectype})

This is the deflated version of the Newton Solver. It penalises the roots saved in defOp.roots

source

newtonFold

newtonHopf

PseudoArcLengthContinuation.continuation — Function.

continuation(F::Function, J, u0, p0::Real, contParams::ContinuationPar; plot = false, printsolution = norm, plotsolution::Function = (x;kwargs...)->nothing, finaliseSolution::Function = (x, y)-> nothing, linearalgo   = :bordered, verbosity = 2)

Compute the continuation curve associated to the functional F and its jacobian J. The parameters are as follows

F = (x, p) -> F(x, p) where p is the parameter for the continuation

J = (x, p) -> d_xF(x, p) its associated jacobian

u0 initial guess

contParams parameters for continuation, with type ContinuationPar

plot = false whether to plot the solution while computing

printsolution = norm function used to plot in the continuation curve, e.g. norm or x -> x[1]

plotsolution::Function = (x; kwargs...)->nothing function implementing the plotting of the solution.

finaliseSolution::Function = (z, tau, step, contResult) -> nothing Function called at the end of each continuation step

linearalgo = :bordered

verbosity controls the amount of information printed during the continuation process.

The function outputs

contres::ContResult structure which contains the computed branch

u::BorderedVector the last solution computed on the branch

Method

Bordered system of equations

The pseudo arclength continuation method solve the equation $F(x,p) = 0$ (or dimension N) together with the pseudo-arclength consraint $N(x, p) = \frac{\theta}{length(u)} \langle x - x_0, \tau_0\rangle + (1 - \theta)\cdot(p - p_0)\cdot dp_0 - ds = 0$. In practice, the curve is parametrised by $s$ so that $(x(s),p(s))$ is a curve of solution to $F(x,p)$. This formulation allows to pass turning points (where the implicit theorem fails). In the previous formula, $(x_0, p_0)$ is a solution for a given $s_0$, $(\tau_0, dp_0)$ is the tangent to the curve at $s_0$. Hence, to compute the solution curve, we solve an equation of dimension N+1 which is called a Bordered system.

Parameter `theta`

The parameter theta in the type ContinuationParis very important. It should be tuned for the continuation to work properly especially in the case of large problems in which cases the $\langle x - x_0, \tau_0\rangle$ component in the constraint might be favoured to much.

The parameter ds is adjusted internally depending on the number of Newton iterations and other factors. See the function stepSizeControl for more information. An important parameter to adjust the magnitude of this adaptation is the parameter a in the type ContinuationPar.

Algorithm

The algorithm works as follows:

Start from a known solution $(x_0,p_0,\tau_0,dp_0)$

Predictor set $(x_1,p_1) = (x_0,p_0) + ds\cdot (\tau_0,dp_0)$

Corrector solve $F(x,p)=0,\ N(x,p)=0$ with a (Bordered) Newton Solver.

New tangent Compute $(\tau_1,dp_1)$, set $(x_0,p_0,\tau_0,dp_0)=(x_1,p_1,\tau_1,dp_1)$ and return to step 2

Natural continuation

We speak of natural continuation when we do not consider the constraint $N(x,p)=0$. Knowing $(x_0,p_0)$, we use $x_0$ as a guess for solving $F(x,p_1)=0$ with $p_1$ close to $p_0$. Again, this will fail at Turning Point but it can be faster to compute than the constrained case. This is set by the field natural in the type ContinuationPar`

Tangent computation (step 4)

There are various ways to compute $(\tau_1,p_1)$. The first one is called secant and is parametrised by the field secant in the type ContinuationPar. It is computed by $(\tau_1,p_1) = (z_1,p_1) - (z_0,p_0)$ and normalised by the norm $\|u,p\|^2_\theta = \frac{\theta}{length(u)} \langle u,u\rangle + (1 - \theta)\cdot p^2$. If secant is set to false, another method is use computing $(\tau_1,p_1)$ by solving a bordered linear system, see the function getTangentBordered for more information.

Bordered linear solver

When solving the Bordered system $F(x,p) = 0,\ N(x, p)=0$, one faces the issue of solving the Bordered linear system $\begin{bmatrix} J & a ; b^T & c\end{bmatrix}\begin{bmatrix}X ; y\end{bmatrix} =\begin{bmatrix}R ; n\end{bmatrix}$. This can be solved in many ways via bordering (which requires to Jacobian inverses) or by forming the bordered matrix (which works well for sparse matrices). The choice of method is set by the argument linearalgo. Have a look at the function linearBorderedSolver for more information.

source

continuationFold

continuationHopf