Fold / Hopf Continuation

For this to work, it is important to have an analytical expression for the jacobian. See the tutorial Temperature model for more details.

The case of the Fold point

Once a Fold point has been detected after a call to br, _ = continuation(...), it can be refined using newton iterations. We have implemented a Minimally Augmented formulation. A simplified interface is provided.

Newton refinement

Let us say that ind_fold is the index in br.bifpoint of a Fold point. This guess can be refined by calling the following simplified interface. More precisions are provided below for an advanced usage.

outfold, hist, flag = @time newtonFold(
					(x, p) ->   F(x, p),
					(x, p) -> Jac(x, p),
					br, ind_fold,
					opt_newton)

It is important to note that for improved performances, a function implementing the expression of the hessian should be provided. This is by far the fastest. Reader interested in this advanced usage should look at the code example/chan.jl of the tutorial Temperature model. Although it is a simple problem, many different use case are shown in a simple setting. See also newtonFold.

The case of the Hopf point

One a Hopf point have been detected after a call to br, _ = continuation(...), it can be refined using newton iterations. We have implemented a Minimally Augmented formulation. A simplified interface is provided as for the Fold case.

Newton refinement

Let us say that ind_hopf is the index in br.bifpoint of a Hopf point. This guess can be refined by calling the simplified interface. More precisions are provided below for an advanced usage. See also newtonHopf.

outfold, hist, flag = @time newtonHopf(
					(x, p) ->   F(x, p),
					(x, p) -> Jac(x, p),
					br, ind_hopf,
					opt_newton)

Methods

PseudoArcLengthContinuation.newtonFold — Function

`newtonFold(F, J, foldpointguess::BorderedArray{vectype, T}, eigenvec, options::NewtonPar; Jt = nothing, d2F = nothing, normN = norm)`

This function turns an initial guess for a Fold point into a solution to the Fold problem based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is the parameter associated to the Fold point

dF = (x, p) -> d_xF(x, p) associated jacobian

foldpointguess initial guess (x0, p0) for the Fold point. It should be a BorderedArray as given by the function FoldPoint

eigenvec guess for the 0 eigenvector

options::NewtonPar

Optional arguments:

Jt = (x, p) -> transpose(d_xF(x, p)) associated jacobian transpose, it should be implemented in an efficient manner. For matrix-free methods, transpose is not readily available and the user must provide a dedicated method.

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) a bilinear operator representing the hessian of F. It has to provide an expression for d2F(x,p)[v1,v2].

normN = norm

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

Simplified call to refine an initial guess for a Fold point. More precisely, the call is as follows

`newtonFold(F, J, br::ContResult, index::Int64, options::NewtonPar; Jt = nothing, d2F = nothing)`

where the optional argument Jt is the jacobian transpose and the Hessian is d2F. The parameters / options are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.newtonHopf — Function

`newtonHopf(F, J, hopfpointguess::BorderedArray{vectypeR, T}, eigenvec, eigenvec_ad, options::NewtonPar; Jt = nothing, d2F = nothing, normN = norm)`

This function turns an initial guess for a Hopf point into a solution to the Hopf problem based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is the parameter associated to the Hopf point

J = (x, p) -> d_xF(x, p) associated jacobian

hopfpointguess initial guess (x0, p0) for the Hopf point. It should a BorderedArray as given by the function HopfPoint.

eigenvec guess for the iω eigenvector

eigenvec_ad guess for the -iω eigenvector

options::NewtonPar

Optional arguments:

Jt = (x, p) -> transpose(d_xF(x, p)) associated jacobian transpose

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) a bilinear operator representing the hessian of F. It has to provide an expression for d2F(x,p)[v1, v2].

normN = norm

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

Simplified call to refine an initial guess for a Hopf point. More precisely, the call is as follows

`newtonHopf(F, J, br::ContResult, index::Int64, options; Jt = nothing, d2F = nothing, normN = norm)`

where the optional argument Jt is the jacobian transpose and the Hessian is d2F. The parameters are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Eigenvectors

This simplified call has been written when the eigenvectors are organised in a 2d Array evec where evec[:,2] is the second eigenvector in the list.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.continuationFold — Function

Codim 2 continuation of Fold points. This function turns an initial guess for a Fold point into a curve of Fold points based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p1, p2) -> F(x, p1, p2) where p is the parameter associated to the Fold point

J = (x, p1, p2) -> d_xF(x, p1, p2) associated jacobian

foldpointguess initial guess (x0, p10) for the Fold point. It should be a BorderedArray as given by the function FoldPoint

p2 parameter p2 for which foldpointguess is a good guess

eigenvec guess for the 0 eigenvector at p1_0

options::NewtonPar

Optional arguments:

Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) associated jacobian transpose

d2F = p2 -> ((x, p1, v1, v2) -> d2F(x, p1, p2, v1, v2)) this is the hessian of F computed at (x, p1, p2) and evaluated at (v1, v2).

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

Simplified call for continuation of Fold point. More precisely, the call is as follows continuationFold(F, J, br::ContResult, index::Int64, options; Jt = nothing, d2F = p2 -> nothing) where the parameters are as for continuationFold except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.continuationHopf — Function

codim 2 continuation of Hopf points. This function turns an initial guess for a Hopf point into a curve of Hopf points based on a Minimally Augmented formulation. The arguments are as follows

(x, p1, p2)-> F(x, p1, p2) where p is the parameter associated to the hopf point

J = (x, p1, p2)-> d_xF(x, p1, p2) associated jacobian

hopfpointguess initial guess (x0, p10) for the Hopf point. It should be a Vector or a BorderedArray

p2 parameter p2 for which hopfpointguess is a good guess

eigenvec guess for the iω eigenvector at p1_0

eigenvec_ad guess for the -iω eigenvector at p1_0

options::NewtonPar

Optional arguments:

Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) associated jacobian transpose

d2F = p2 -> ((x, p1, v1, v2) -> d2F(x, p1, p2, v1, v2)) this is the hessian of F computed at (x, p1, p2) and evaluated at (v1, v2).

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

Simplified call for continuation of Hopf point. More precisely, the call is as follows continuationHopf(F, J, br::ContResult, index::Int64, options; Jt = nothing, d2F = p2 -> nothing) where the parameters are as for continuationHopf except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Simplified calls are also provided but at the cost of using finite differences.

Eigenvectors

This simplified call has been written when the eigenvectors are organised in a 2d Array evec where evec[:,2] is the second eigenvector in the list.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6