Fold / Hopf Continuation

For this to work best, it is necessary to have an analytical expression for the jacobian. See the tutorial Temperature model (Simplest example) for more details.

In this page, we explain how to perform continuation of Fold / Hopf points and detect the associated bifurcations.

List of detected bifurcation points

Bifurcation	index used

Bogdanov-Takens	bt

Bautin	gh

Cusp	cusp

In a nutshell, all you have to do (see below) is to call continuation(F, J, br, ind_bif) to continue the bifurcation point stored in br.specialpoint[ind_bif] and set proper options.

Fold continuation

The continuation of Fold bifurcation points is based on a Minimally Augmented[Govaerts] formulation which is an efficient way to detect singularities. The continuation of Fold points is based on the formulation $G(u,p) = (F(u,p), g(u,p))\in\mathbb R^{n+1}$ where the test function $g$ is solution of

\[\left[\begin{array}{cc} dF(u,p) & w \\ v^{\top} & 0 \end{array}\right]\left[\begin{array}{c} r \\ g(u,p) \end{array}\right]=\left[\begin{array}{c}0_{n} \\1\end{array}\right]\quad\quad (M_f)\]

where $w,v$ are chosen in order to have a non-singular matrix $(M_f)$. More precisely, $v$ (resp. $w$) should be close to a null vector of dF(u,p) (resp. dF(u,p)'). During continuation, the vectors $w,v$ are updated so that the matrix $(M_f)$ remains non-singular ; this is controlled with the argument updateMinAugEveryStep (see below).

note that there are very simplified calls for this, see Newton refinement below. In particular, you don't need to set up the Fold Minimally Augmented problem yourself. This is done in the background.

Linear Method

You can pass the bordered linear solver to solve $(M_f)$ using the option bdlinsolver (see below). Note that the choice bdlinsolver = BorderingBLS() can lead to singular systems. Indeed, in this case, $(M_f)$ is solved by inverting dF(u,p) which is singular at Fold points.

Hopf continuation

The continuation of Fold bifurcation points is based on a Minimally Augmented (see [Govaerts] p. 87) formulation which is an efficient way to detect singularities. The continuation of Hopf points is based on the formulation $G(u,\omega,p) = (F(u,\omega,p), g(u,\omega,p))\in\mathbb R^{n+2}$ where the test function $g$ is solution of

\[\left[\begin{array}{cc} dF(u,p)-i\omega I_n & w \\ v^{\top} & 0 \end{array}\right]\left[\begin{array}{c} r \\ g(u,\omega,p) \end{array}\right]=\left[\begin{array}{c} 0_{n} \\ 1 \end{array}\right]\quad\quad (M_h)\]

where $w,v$ are chosen in order to have a non-singular matrix $(M_h)$. More precisely, $w$ (resp. $v$) should be a left (resp. right) approximate null vector of $dF(u,p)-i\omega I_n$. During continuation, the vectors $w,v$ are updated so that the matrix $(M_h)$ remains non-singular ; this is controlled with the argument updateMinAugEveryStep (see below).

note that there are very simplified calls to this, see Newton refinement below. In particular, you don't need to set up the Hopf Minimally Augmented problem yourself. This is done in the background.

Linear Method

You can pass the bordered linear solver to solve $(M_h)$ using the option bdlinsolver (see below). Note that the choice bdlinsolver = BorderingBLS() can lead to singular systems. Indeed, in this case, $(M_h)$ is solved by inverting dF(u,p) which is singular at Fold points.

Newton refinement

Once a Fold/Hopf point has been detected after a call to br, = continuation(...), it can be refined using newton iterations. Let us say that ind_bif is the index in br.specialpoint of a Fold/Hopf point. This guess can be refined as follows:

outfold, hist, flag =  newton(F, J, br::AbstractBranchResult, ind_bif::Int; 
	issymmetric = false, Jᵗ = nothing, d2F = nothing, 
	normN = norm, options = br.contparams.newtonOptions, 
	bdlinsolver = BorderingBLS(options.linsolver),
	startWithEigen = false, kwargs...)

where par is the set of parameters used in the call to continuation to compute br. For the options parameters, we refer to Newton.

It is important to note that for improved performances, a function implementing the expression of the hessian should be provided. This is by far the fastest. Reader interested in this advanced usage should look at the code example/chan.jl of the tutorial Temperature model (Simplest example).

Codim 2 continuation

To compute the codim 2 curve of Fold/Hopf points, one can call continuation with the following options

BifurcationKit.continuation — Function

continuation(F, J, br, ind_bif, lens2)
continuation(F, J, br, ind_bif, lens2, options_cont; startWithEigen, issymmetric, Jᵗ, d2F, d3F, detectCodim2Bifurcation, kwargs...)

codim 2 continuation of Fold / Hopf points. This function turns an initial guess for a Fold/Hopf point into a curve of Fold/Hopf points based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is a set of parameters

J = (x, p) -> d_xF(x, p) associated jacobian

br results returned after a call to continuation

ind_bif bifurcation index in br

lens2 parameters used for the vector field

options_cont = br.contparams arguments to be passed to the regular continuation

Optional arguments:

issymmetric whether the Jacobian is Symmetric (for Fold)

Jᵗ = (x, p) -> transpose(d_xF(x, p)) associated jacobian transpose

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) this is the hessian of F computed at (x, p) and evaluated at (v1, v2). This helps solving the linear problem associated to the minimally augmented formulation.

d3F = (x, p, v1, v2, v3) -> d3F(x, p, v1, v2, v3) this is the third derivative of F computed at (x, p) and evaluated at (v1, v2, v3). This is used to detect Bautin bifurcation.

bdlinsolver bordered linear solver for the constraint equation

updateMinAugEveryStep update vectors a,b in Minimally Formulation every updateMinAugEveryStep steps

startWithEigen = false whether to start the Minimally Augmented problem with information from eigen elements

detectCodim2Bifurcation ∈ {0,1,2} whether to detect Bogdanov-Takens, Bautin and Cusp. If equals 1 non precise detection is used. If equals 2, a bisection method is used to locate the bifurcations.

kwargs keywords arguments to be passed to the regular continuation

where the parameters are as above except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of Hopf point in br you want to refine.

Jacobian transpose

The adjoint of the jacobian J is computed internally when Jᵗ = nothing by using transpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jᵗ = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian would be computed twice!

ODE problems

For ODE problems, it is more efficient to pass the Bordered Linear Solver using the option bdlinsolver = MatrixBLS()

where the options are as above except with have an additional parameter axis lens2 which is used to locate the bifurcation points.

See Temperature model (Simplest example) for an example of use.

Advanced use

Here, we expose the solvers that are used to perform newton refinement or codim 2 continuation in case the above methods fails. This is useful in case it is too involved to expose the linear solver options. An example of advanced use is the continuation of Folds of periodic orbits, see Continuation of Fold of periodic orbits.

BifurcationKit.newtonFold — Function

newtonFold(F, J, foldpointguess, par, lens, eigenvec, eigenvec_ad, options; normN, issymmetric, Jᵗ, d2F, bdlinsolver, kwargs...)

This function turns an initial guess for a Fold point into a solution to the Fold problem based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is a set of parameters.

dF = (x, p) -> d_xF(x, p) associated jacobian

foldpointguess initial guess (x0, p0) for the Fold point. It should be a BorderedArray as returned by the function FoldPoint

par parameters used for the vector field

lens parameter axis used to locate the Fold point.

eigenvec guess for the 0 eigenvector

eigenvec_ad guess for the 0 adjoint eigenvector

options::NewtonPar options for the Newton-Krylov algorithm, see NewtonPar.

Optional arguments:

issymmetric whether the Jacobian is Symmetric

Jᵗ = (x, p) -> transpose(d_xF(x, p)) jacobian adjoint, it should be implemented in an efficient manner. For matrix-free methods, transpose is not readily available and the user must provide a dedicated method. In the case of sparse based jacobian, Jᵗ should not be passed as it is computed internally more efficiently, i.e. it avoids recomputing the jacobian as it would be if you pass Jᵗ = (x, p) -> transpose(dF(x, p))

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) a bilinear operator representing the hessian of F. It has to provide an expression for d2F(x, p)[v1, v2].

normN = norm

bdlinsolver bordered linear solver for the constraint equation

kwargs keywords arguments to be passed to the regular Newton-Krylov solver

Simplified call

Simplified call to refine an initial guess for a Fold point. More precisely, the call is as follows

newtonFold(F, J, br::AbstractBranchResult, ind_fold::Int, lens::Lens; options = br.contparams.newtonOptions, kwargs...)

where the optional argument Jᵗ is the jacobian transpose and the Hessian is d2F. The parameters / options are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine. You can pass newton parameters different from the ones stored in br by using the argument options.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jᵗ = nothing by using transpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jᵗ = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

ODE problems

For ODE problems, it is more efficient to pass the Bordered Linear Solver using the option bdlinsolver = MatrixBLS()

BifurcationKit.newtonHopf — Function

newtonHopf(F, J, hopfpointguess, par, lens, eigenvec, eigenvec_ad, options; Jᵗ, d2F, normN, bdlinsolver, kwargs...)

This function turns an initial guess for a Hopf point into a solution to the Hopf problem based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is a set of parameters.

dF = (x, p) -> d_xF(x, p) associated jacobian

hopfpointguess initial guess (x0, p0) for the Hopf point. It should a BorderedArray as returned by the function HopfPoint.

par parameters used for the vector field

lens parameter axis used to locate the Hopf point.

eigenvec guess for the iω eigenvector

eigenvec_ad guess for the -iω eigenvector

options::NewtonPar options for the Newton-Krylov algorithm, see NewtonPar.

Optional arguments:

Jᵗ = (x, p) -> transpose(d_xF(x, p)) jacobian adjoint, it should be implemented in an efficient manner. For matrix-free methods, transpose is not readily available and the user must provide a dedicated method. In the case of sparse based jacobian, Jᵗ should not be passed as it is computed internally more efficiently, i.e. it avoids recomputing the jacobian as it would be if you pass Jᵗ = (x, p) -> transpose(dF(x, p))

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) a bilinear operator representing the hessian of F. It has to provide an expression for d2F(x, p)[v1, v2].

normN = norm

bdlinsolver bordered linear solver for the constraint equation

kwargs keywords arguments to be passed to the regular Newton-Krylov solver

Simplified call:

Simplified call to refine an initial guess for a Hopf point. More precisely, the call is as follows

newtonHopf(F, J, br::AbstractBranchResult, ind_hopf::Int, lens::Lens; Jᵗ = nothing, d2F = nothing, normN = norm, options = br.contparams.newtonOptions, kwargs...)

where the optional argument Jᵗ is the jacobian transpose and the Hessian is d2F. The parameters / options are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine. You can pass newton parameters different from the ones stored in br by using the argument options.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jᵗ = nothing by using transpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jᵗ = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

ODE problems

For ODE problems, it is more efficient to pass the Bordered Linear Solver using the option bdlinsolver = MatrixBLS()

BifurcationKit.continuationFold — Function

continuationFold(F, J, foldpointguess, par, lens1, lens2, eigenvec, eigenvec_ad, options_cont; normC, updateMinAugEveryStep, issymmetric, Jᵗ, d2F, bdlinsolver, kwargs...)

Codim 2 continuation of Fold points. This function turns an initial guess for a Fold point into a curve of Fold points based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is a set of parameters

J = (x, p) -> d_xF(x, p) associated jacobian

foldpointguess initial guess (x0, p10) for the Fold point. It should be a BorderedArray as returned by the function FoldPoint

par set of parameters

lens1 parameter axis for parameter 1

lens2 parameter axis for parameter 2

eigenvec guess for the 0 eigenvector at p1_0

eigenvec_ad guess for the 0 adjoint eigenvector

options_cont arguments to be passed to the regular continuation

Optional arguments:

issymmetric whether the Jacobian is Symmetric

Jᵗ = (x, p) -> transpose(d_xF(x, p)) associated jacobian transpose

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) this is the hessian of F computed at (x, p) and evaluated at (v1, v2).

bdlinsolver bordered linear solver for the constraint equation

updateMinAugEveryStep update vectors a, b in Minimally Formulation every updateMinAugEveryStep steps

kwargs keywords arguments to be passed to the regular continuation

Simplified call

The call is as follows

continuationFold(F, J, br::AbstractBranchResult, ind_fold::Int64, lens2::Lens, options_cont::ContinuationPar ; kwargs...)

where the parameters are as above except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of Fold point in br you want to continue.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jᵗ = nothing by using transpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jᵗ = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian would be computed twice!

ODE problems

For ODE problems, it is more efficient to pass the Bordered Linear Solver using the option bdlinsolver = MatrixBLS()

Detection of Bogdanov-Takens and Cusp bifurcations

In order to trigger the detection, pass detectEvent = 1,2 in options_cont.

BifurcationKit.continuationHopf — Function

codim 2 continuation of Hopf points. This function turns an initial guess for a Hopf point into a curve of Hopf points based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is a set of parameters

J = (x, p) -> d_xF(x, p) associated jacobian

hopfpointguess initial guess (x0, p10) for the Hopf point. It should be a Vector or a BorderedArray

par set of parameters

lens1 parameter axis for parameter 1

lens2 parameter axis for parameter 2

eigenvec guess for the iω eigenvector at p1_0

eigenvec_ad guess for the -iω eigenvector at p1_0

options_cont keywords arguments to be passed to the regular continuation

Optional arguments:

Jᵗ = (x, p) -> adjoint(d_xF(x, p)) associated jacobian adjoint

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) this is the hessian of F computed at (x, p) and evaluated at (v1, v2). This helps solving the linear problem associated to the Hopf minimally augmented formulation.

d3F = (x, p, v1, v2, v3) -> d3F(x, p, v1, v2, v3) this is the third derivative of F computed at (x, p) and evaluated at (v1, v2, v3). This is used to detect Bautin bifurcation.

bdlinsolver bordered linear solver for the constraint equation

updateMinAugEveryStep update vectors a,b in Minimally Formulation every updateMinAugEveryStep steps

kwargs keywords arguments to be passed to the regular continuation

Simplified call:

The call is as follows

continuationHopf(F, J, br::AbstractBranchResult, ind_hopf::Int, lens2::Lens, options_cont::ContinuationPar ;  kwargs...)

where the parameters are as above except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of Hopf point in br you want to refine.

ODE problems

For ODE problems, it is more efficient to pass the Bordered Linear Solver using the option bdlinsolver = MatrixBLS()

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jᵗ = nothing by using transpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jᵗ = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian would be computed twice!

Detection of Bogdanov-Takens and Bautin bifurcations

In order to trigger the detection, pass detectEvent = 1,2 in options_cont. Note that you need to provide d3F.

References

Govaerts

Govaerts, Willy J. F. Numerical Methods for Bifurcations of Dynamical Equilibria. Philadelphia, Pa: Society for Industrial and Applied Mathematics, 2000.