Library

Structs

PseudoArcLengthContinuation.NewtonPar — Type

options = NewtonPar(tol = 1e-4,...)

Returns a variable containing parameters to affect the newton algorithm when solving F(x) = 0.

Arguments (with default values):

tol = 1e-10: absolute tolerance for F(x)

maxIter = 50: number of Newton iterations

verbose = false: display Newton iterations?

linsolver = DefaultLS(): linear solver, must be <: AbstractLinearSolver

eigsolver = DefaultEig(): eigen solver, must be <: AbstractEigenSolver

Arguments only used in newtonPALC

linesearch = false: use line search algorithm

alpha = 1.0: alpha (damping) parameter for line search algorithm

almin = 0.001: minimal vslue of the damping alpha

Mutating

For performance reasons, we decided to use an immutable structure to hold the parameters. One can use the package Setfield.jl to drastically simplify the mutation of different fields. See tutorials for more examples.

PseudoArcLengthContinuation.ContinuationPar — Type

options = ContinuationPar(dsmin = 1e-4,...)

Returns a variable containing parameters to affect the continuation algorithm when solving F(x,p) = 0.

Arguments

dsmin, dsmax are the minimum, maximum arclength allowed value. It controls the density of points in the computed branch of solutions.

ds is the initial arclength.

theta is a parameter in the arclength constraint. It is very important to tune it. See the docs of continuation.

pMin, pMax allowed parameter range for p

maxSteps maximum number of continuation steps

newtonOptions::NewtonPar: options for the Newton algorithm

saveToFile = false: save to file. A name is automatically generated.

saveSolEveryNsteps::Int64 = 0 at which continuation steps do we save the current solution`

plotEveryNsteps = 3

Handling eigen elements

computeEigenValues = false compute eigenvalues / eigenvectors

nev = 3 number of eigenvalues to be computed. It is automatically increased to have at least nev unstable eigenvalues. To be set for proper bifurcation detection. See Detection of bifurcation points for more informations.

saveEigEveryNsteps = 1 record eigen vectors every specified steps. Important for memory limited ressource, e.g. GPU.

saveEigenvectors = true Important for memory limited ressource, e.g. GPU.

Handling bifurcation detection

precisionStability = 1e-10 lower bound on the real part of the eigenvalues to test for stability of equilibria and periodic orbits

detectFold = true detect Fold bifurcations? It is a useful option although the detection of Fold is cheap. Indeed, it may happen that there is a lot of Fold points and this can saturate the memory in memory limited devices (e.g. on GPU)

detectBifurcation::Int ∈ {0, 1, 2} detect bifurcations? If set to 0, bifurcation are not detected. If set to 1, bifurcation are detected along the continuation run, but not located precisely. If set to 2, a bisection algorithm is used to locate the bifurcations (slower). The possibility to switch off detection is a useful option. Indeed, it may happen that there are a lot of bifurcation points and this can saturate the memory in memory limited devices (e.g. on GPU)

dsminBisection dsmin for the bisection algorithm when locating bifurcation points

nInversion number of sign inversions in bisection algorithm

maxBisectionSteps maximum number of bisection steps

Handling ds adaptation (see continuation for more information)

a = 0.5 aggressiveness factor. It is used to adapt ds in order to have a number of newton iteration per continuation step roughly the same. The higher a is, the larger the step size ds is changed at each continuation step.

thetaMin = 1.0e-3 minimum value of theta

doArcLengthScaling trigger further adaptation of theta

Misc

finDiffEps::T = 1e-9 ε used in finite differences computations

Mutating

For performance reasons, we decided to use an immutable structure to hold the parameters. One can use the package Setfield.jl to drastically simplify the mutation of different fields. See tutorials for more examples.

PseudoArcLengthContinuation.PrecPartialSchurKrylovKit — Function

PrecPartialSchurKrylovKit(J, x0, nev, which = :LM; krylovdim = max(2nev, 20), verbosity = 0)

Builds a preconditioner based on deflation of nev eigenvalues chosen according to which. A partial Schur decomposition is build (Matrix-Free) from which a projection is built. The package KrylovKit.jl is used to compute the partial Schur decomposition. The options are similar to the ones of EigKrylovKit().

PseudoArcLengthContinuation.PrecPartialSchurArnoldiMethod — Function

PrecPartialSchurArnoldiMethod(J, N, nev, which = LM(); tol = 1e-9, kwargs...)

Builds a preconditioner based on deflation of nev eigenvalues chosen according to which. A partial Schur decomposition is build (Matrix-Free) from which a projection is built. The package ArnoldiMethod.jl is used to compute the partial Schur decomposition. See the package ArnoldiMethod.jl for how to pass the proper options.

PseudoArcLengthContinuation.ContResult — Type

ContResult{T, Eigentype, Vectype, Biftype}

Structure which holds the results after a call to continuation.

Arguments

branch::VectorOfArray holds the low-dimensional information about the branch. More precisely, branch[:,i] contains the following information (param, printSolution(u, param), Newton iterations, ds, i) for each continuation step i.

n_unstable::Vector{Int64} a vector holding the number of eigenvalues with positive real part for each continuation step (to detect stationary bifurcation)

n_imag::Vector{Int64} a vector holding the number of eigenvalues with positive real part and non zero imaginary part for each continuation step (to detect Hopf bifurcation)

stability::Vector{Bool} a vector holding the stability of the computed solution for each continuation step

bifpoint::Vector{Biftype} a vector holding the set of bifurcation points detected during the computation of the branch. Each entry of the vector contains a tuple (type, idx, param = T(0.), norm, printsol, u, tau, ind_bif, step, status) where step is the continuation step at which the bifurcation occurs, ind_bif is the eigenvalue index responsible for the bifurcation (if applicable) and idx is the index in eig (see above) for which the bifurcation occurs.

foldpoint::Vector{Biftype} a vector holding the set of fold points detected during the computation of the branch.

eig::Vector contains for each continuation step the eigen elements.

PseudoArcLengthContinuation.DeflationOperator — Type

DeflationOperator. It is used to describe the following situation. Assume you want to solve F(x)=0 with a Newton algorithm but you want to avoid the process to return some already known solutions $roots_i$. You can use DeflationOperator to define a function M(u) used to find, with Newton iterations, the zeros of the following function $F(u) / Π_i(dot(u - roots_i, u - roots_i)^{p} + shift) := F(u) / M(u)$. The fields of the struct DeflationOperator are as follows:

power p

dot function, this function has to be bilinear and symmetric for the linear solver

shift

roots

The deflation operator is is $M(u) = \frac{1}{\prod_{i=1}^{n_{roots}}(shift + norm(u-roots_i)^p)}$ where $nrm(u) = dot(u,u)$.

Given defOp::DeflationOperator, one can access its roots as defOp[n] as a shortcut for defOp.roots[n]. Also, one can add (resp.remove) a new root by using push!(defOp, newroot) (resp. pop!(defOp)). Finally length(defOp) is a shortcut for length(defOp.roots)

PseudoArcLengthContinuation.DeflatedProblem — Type

pb = DeflatedProblem(F, J, M::DeflationOperator)

This creates a problem encoded a deflated problem $M(u) \cdot F(u) = 0$ where M is a DeflationOperator which encodes the penalization term. J is the jacobian of J. Can be used to call newton and continuation.

PseudoArcLengthContinuation.PeriodicOrbitTrapProblem — Type

pb = PeriodicOrbitTrapProblem(F, J, ϕ, xπ, M::Int)

This composite type implements Finite Differences based on a Trapezoidal rule to locate periodic orbits. The arguments are as follows

F vector field

J jacobian of F

d2F = nothing Hessian of F (optional), useful for continuation of Fold of periodic orbits

ϕ used to set a section for the phase constraint equation

xπ used in the section for the phase constraint equation

M::Int number of time slices

linsolver: = DefaultLS() linear solver for each time slice, i.e. to solve J⋅sol = rhs. This is only used for the computation of the Floquet multipliers.

isinplace::Bool whether F and J are inplace functions (Experimental). In this case, the functions F and J must have the following definitions (o, x) -> F(o, x) and (o, x, dx) -> J(o, x, dx).

ongpu::Bool whether the computation takes place on the gpu (Experimental)

You can then call pb(orbitguess) to compute the functional on a orbitguess. Note that orbitguess must be of size M * N + 1 where N is the number of unknowns in the state space and orbitguess[M*N+1] is an estimate of the period of the limit cycle.

The scheme is as follows, one look for T = x[end] and

$\left(x_{i} - x_{i-1}\right) - \frac{h}{2} \left(F(x_{i}) + F(x_{i-1})\right) = 0,\ i=1,\cdots,m-1$

with u_{0} = u_{m-1} and the periodicity condition u_{m} - u_{1} = 0 and

where h = T/M. Finally, the phase of the periodic orbit is constrained by using a section

$\langle x[1] - x_\pi, \phi\rangle=0.$

A functional, hereby called G, encodes this problem. The following methods are available

pb(orbitguess) evaluates the functional G on orbitguess

pb(orbitguess, du) evaluates the jacobian dG(orbitguess).du functional at orbitguess on du

pb(Val(:JacFullSparse), orbitguess) return the sparse matrix of the jacobian dG(orbitguess) at orbitguess without the constraints. It is called A_γ in the docs.

pb(Val(:JacCyclicSparse), orbitguess) return the sparse cyclic matrix Jc (see the docs) of the jacobian dG(orbitguess) at orbitguess

pb(Val(:BlockDiagSparse), orbitguess) return the diagonal of the sparse matrix of the jacobian dG(orbitguess) at orbitguess. This allows to design Jacobi preconditioner. Use blockdiag.

GPU call

For these methods to work on the GPU, for example with CuArrays in mode allowscalar(false), we face the issue that the function extractPeriodFDTrap won't be well defined because it is a scalar operation. One may have to redefine it like extractPeriodFDTrap(x::CuArray) = x[end:end] or something else. Also, note that you must pass the option ongpu = true for the functional to be evaluated efficiently on the gpu.

PseudoArcLengthContinuation.Flow — Type

fl = Flow(F, flow, dflow)

This composite type encapsulates:

the vector field x -> F(x) associated to a Cauchy problem,

the flow (or semigroup) associated to the Cauchy problem (x, t) -> flow(x, t). Only the last time point must be returned.

the flow (or semigroup) associated to the Cauchy problem (x, t) -> flow(x, t). The whole solution on the time interval (0,t) must be returned. This is not strictly necessary to provide this.

the differential dflow of the flow w.r.t. x, (x, dx, t) -> dflow(x, dx, t). One important thing is that we require dflow(x, dx, t) to return 2 vectors: the first is flow(x, t) and the second is the value of the derivative of the flow as the second vector.

There are some simple constructors for which you only have to pass a prob::ODEProblem from DifferentialEquations.jl and an ODE time stepper like Tsit5(). Hence, you can do for example

fl = Flow(F, prob, Tsit5(); kwargs...)

If your vector field depends on parameters p, you can define a Flow using

fl = Flow(F, p, prob, Tsit5(); kwargs...)

Finally, you can pass two ODEProblem where the second one is used to compute the variational equation:

fl = Flow(F, p, prob1::ODEProblem, alg1, prob2::ODEProblem, alg2; kwargs...)

PseudoArcLengthContinuation.ShootingProblem — Type

pb = ShootingProblem(flow::Flow, ds, section)

This composite type implements the Standard Simple / Multiple Standard Shooting method to locate periodic orbits. The arguments are as follows

flow::Flow: implements the flow of the Cauchy problem though the structure Flow.

ds: vector of time differences for each shooting. Its length is written M. If M==1, then the simple shooting is implemented and the multiple one otherwise.

section: implements a phase condition. The evaluation section(x) must return a scalar number where x is a guess for the periodic orbit. Note that the period T of the guess x is always included either as the last component of T = x[end] or as T = x.p. The type of x depends on what is passed to the newton solver.

You can then call pb(orbitguess) to apply the functional to a guess. Note that orbitguess must be of size M * N + 1 where N is the number of unknowns of the state space and orbitguess[M * N + 1] is an estimate of the period T of the limit cycle. This form of guess is convenient for the use of the linear solvers in IterativeSolvers.jl (for example) which accepts only AbstractVectors. Another accepted guess is of the form BorderedArray(guess, T) where guess[i] is the state of the orbit at the ith time slice. This last form allows for non-vector state space which can be convenient for 2d problems for example.

A functional, hereby called G, encodes the shooting problem. For example, the following methods are available:

pb(orbitguess) evaluates the functional G on orbitguess

pb(orbitguess, du) evaluates the jacobian dG(orbitguess).du functional at orbitguess on du

Simplified constructors

A simpler way to build a functional is to use

pb = ShootingProblem(F, p, prob::ODEProblem, alg, M::Int, section; kwargs...)

or

pb = ShootingProblem(F, p, prob::ODEProblem, alg, ds, section; kwargs...)

where F is the vector field, p is a parameter (to be passed to the vector field and the flow), prob is an ODEProblem which is used to create a flow using the ODE solver alg (for example Tsit5()). Finally, the arguments kwargs are passed to the ODE solver defining the flow. Look at DifferentialEquations.jl for more information. Note that, in this case, the derivative of the flow is computed internally using Finite Differences.

The other way is an elaboration of the previous one

pb = ShootingProblem(F, p, prob1::ODEProblem, alg1, prob2::ODEProblem, alg2, M::Int, section; kwargs...)

or

pb = ShootingProblem(F, p, prob1::ODEProblem, alg1, prob2::ODEProblem, alg2, ds, section; kwargs...)

where we supply now two ODEProblem. The first one prob1, is used to define the flow associated to F while the second one is a problem associated to the derivative of the flow. Hence, prob2 should implement the following vector field $\tilde F(x,\phi) = (F(x),dF(x)\cdot\phi)$.

PseudoArcLengthContinuation.PoincareShootingProblem — Type

pb = PoincareShootingProblem(flow::Flow, M, section)

This composite type implements the Poincaré Shooting method to locate periodic orbits, basically using Poincaré return maps. The arguments are as follows

flow::Flow: implements the flow of the Cauchy problem though the structure Flow.

M: the number of return maps. If M==1, then the simple shooting is implemented and the multiple one otherwise.

section: implements a Poincaré section condition. The evaluation section(x) must return a scalar number where x is a guess for the periodic orbit when M=1. Otherwise, one must implement a function section(out, x) which populates out with the M hyperplanes.

Simplified constructors

A simpler way is to create a functional is pb = PoincareShootingProblem(F, p, prob::ODEProblem, alg, section; kwargs...) for simple shooting or pb = PoincareShootingProblem(F, p, prob::ODEProblem, alg, M::Int, section; kwargs...) for multiple shooting . Here F is the vector field, p is a parameter (to be passed to the vector and the flow), prob is an ODEProblem which is used to create a flow using the ODE solver alg (for example Tsit5()). Finally, the arguments kwargs are passed to the ODE solver defining the flow. Look at DifferentialEquations.jl for more information.

Another convenient call is pb = PoincareShootingProblem(F, p, prob::ODEProblem, alg, normals::AbstractVector, centers::AbstractVector; δ = 1e-8, kwargs...) where normals (resp. centers) is a list of normals (resp. centers) which define a list of hyperplanes $\Sigma_i$. These hyperplanes are used to define partial Poincaré return maps. δ is a numerical value used for the Matrix-Free Jacobian by finite differences. If set to 0, analytical jacobian is used. See docs for more information.

Computing the functionals

You can then call pb(orbitguess) to apply the functional to a guess. Note that orbitguess must be of size M * N where N is the number of unknowns in the state space and M is the number of Poincaré maps. Another accepted guess is such that guess[i] is the state of the orbit on the ith section. This last form allows for non-vector state space which can be convenient for 2d problems for example.

A functional, hereby called G encodes this shooting problem. For example, the following methods are available:

pb(orbitguess) evaluates the functional G on orbitguess

pb(orbitguess, du) evaluates the jacobian dG(orbitguess).du functional at orbitguess on du

Tip

You can use the function getPeriod(pb::PoincareShootingProblem, sol) to get the period of the solution sol

PseudoArcLengthContinuation.FloquetQaDTrap — Type

floquet = FloquetQaDTrap(eigsolver::AbstractEigenSolver)

This composite type implements the computation of the eigenvalues of the monodromy matrix in the case of periodic orbits problems based on Finite Differences (Trapeze method), also called the Floquet multipliers. The method, dubbed Quick and Dirty (QaD), is not numerically very precise for large / small Floquet exponents. It allows, nevertheless, to detect bifurcations. The arguments are as follows:

eigsolver::AbstractEigenSolver solver used to compute the eigenvalues.

If eigsolver == DefaultEig(), then the monodromy matrix is formed and its eigenvalues are computed. Otherwise, a Matrix-Free version of the monodromy is used.

Floquet multipliers computation

The computation of Floquet multipliers is necessary for the detection of bifurcations of periodic orbits (which is done by analyzing the Floquet exponents obtained from the Floquet multipliers). Hence, the eigensolver eigsolver needs to compute the eigenvalues with largest modulus (and not with largest real part which is their default behavior). This can be done by changing the option which = :LM of eigsolver. Nevertheless, note that for most implemented eigensolvers in the current Package, the proper option is set.

PseudoArcLengthContinuation.FloquetQaDShooting — Type

floquet = FloquetQaDShooting(eigsolver::AbstractEigenSolver)

This composite type implements the computation of the eigenvalues of the monodromy matrix in the case of periodic orbits problems based on the Shooting method, also called the Floquet multipliers. The method, dubbed Quick and Dirty (QaD), is not numerically very precise for large / small Floquet exponents. It allows, nevertheless, to detect bifurcations. The arguments are as follows:

eigsolver::AbstractEigenSolver solver used to compute the eigenvalues.

If eigsolver == DefaultEig(), then the monodromy matrix is formed and its eigenvalues are computed. Otherwise, a Matrix-Free version of the monodromy is used.

Floquet multipliers computation

The computation of Floquet multipliers is necessary for the detection of bifurcations of periodic orbits (which is done by analyzing the Floquet exponents obtained from the Floquet multipliers). Hence, the eigensolver eigsolver needs to compute the eigenvalues with largest modulus (and not with largest real part which is their default behavior). This can be done by changing the option which = :LM of eigsolver. Nevertheless, note that for most implemented eigensolvers in the current Package, the proper option is set.

Newton

PseudoArcLengthContinuation.newton — Function

newton(F, J, x0, options::NewtonPar; normN = norm, callback = (x, f, J, res, iteration, optionsN; kwargs...) -> true, kwargs...)

This is the Newton Solver for F(x) = 0 with Jacobian J and initial guess x0. The function normN allows to specify a norm for the convergence criteria. It is important to set the linear solver options.linsolver properly depending on your problem. This linear solver is used to solve $J(x)u = -F(x)$ in the Newton step. You can for example use linsolver = Default() which is the operator backslash: it works well for Sparse / Dense matrices. See Linear solvers for more informations.

Arguments:

x -> F(x) functional whose zeros are looked for. In particular, it is not inplace,

dF(x) = x -> J(x) compute the jacobian of F at x. It is then passed to options.linsolver. The Jacobian J(x) can be a matrix or an out-of-place function.

x0 initial guess

options variable holding the internal parameters used by the newton method

callback function passed by the user which is called at the end of each iteration. Can be used to update a preconditionner for example. The optionsN will be options passed in order to change the linear / eigen solvers

kwargs arguments passed to the callback. Useful when newton is called from continuation

Simplified calls are provided, for example when J is not passed. It is then computed with finite differences.

Output:

solution:

history of residuals

flag of convergence

number of iterations

function newton(Fhandle, Jhandle, x0::vectype, options:: NewtonPar{T}, defOp::DeflationOperator{T, Tf, vectype}; kwargs...) where {T, Tf, vectype}

This is the deflated version of the Newton Solver for F(x) = 0 with Jacobian J. It penalises the roots saved in defOp.roots. The other arguments are as for newton. See DeflationOperator for more informations.

Simplified calls are provided, for example when J is not passed. It then computed with finite differences.

Output:

solution:

history of residuals

flag of convergence

number of iterations

newton(prob::T, orbitguess, options::NewtonPar; kwargs...) where {T <: AbstractShootingProblem}

This is the Newton Solver for computing a periodic orbit using Shooting method.

Output:

solution

history of residuals

flag of convergence

number of iterations

newton(prob::T, orbitguess, options::NewtonPar, defOp::DeflationOperator; kwargs...) where {T <: AbstractShootingProblem}

This is the deflated Newton Solver for computing a periodic orbit using Shooting method.

Output:

solution

history of residuals

flag of convergence

number of iterations

newton(prob::PeriodicOrbitTrapProblem, orbitguess, options::NewtonPar, linearPO = :BorderedLU; kwargs...)

This is the Newton Solver for computing a periodic orbit using a functional G based on Finite Differences and a Trapezoidal rule.

Arguments:

prob a problem of type PeriodicOrbitTrapProblem encoding the functional G

orbitguess a guess for the periodic orbit where orbitguess[end] is an estimate of the period of the orbit. It should be a vector of size N * M + 1 where M is the number of time slices, N is the dimension of the phase space. This must be compatible with the numbers N,M in prob.

options same as for the regular newton method

linearPO = :BorderedLU. Specify the choice of the linear algorithm, which must belong to [:FullLU, :BorderedLU, :FullMatrixFree, :BorderedMatrixFree]. This is used to select a way of inverting the jacobian dG of the functional G. For :FullLU, we use the default linear solver on a sparse matrix representation of dG. For :BorderedLU, we take advantage of the bordered shape of the linear solver and use LU decomposition to invert dG using a bordered linear solver. This is the default algorithm. For :FullMatrixFree, a matrix free linear solver is used for dG: note that a preconditioner is very likely required here because of the cyclic shape of dG which affects the convergence properties of GMRES. Finally, for :BorderedMatrixFree, a matrix free linear solver is used but for Jc only (see docs): it means that options.linsolver is used to invert Jc. These two Matrix-Free options thus expose different part of the jacobian dG in order to use specific preconditioners. For example, an ILU preconditioner on Jc could remove the constraints in dG and lead to poor convergence. Of course, for these last two methods, a preconditioner is likely to be required.

Output:

solution

history of residuals

flag of convergence

number of iterations

newton(probPO::PeriodicOrbitTrapProblem, orbitguess, options::NewtonPar, defOp::DeflationOperator{T, Tf, vectype}, linearPO = :BorderedLU; kwargs...) where {T, Tf, vectype}

This function is similar to newton(probPO, orbitguess, options, linearPO; kwargs...) except that it uses deflation in order to find periodic orbits different from the one in defOp. For example, it can be used in the vicinity of a Hopf bifurcation to prevent the Newton algorithm from converging to the equilibrium point.

Newton with deflation

PseudoArcLengthContinuation.newton — Method

function newton(Fhandle, Jhandle, x0::vectype, options:: NewtonPar{T}, defOp::DeflationOperator{T, Tf, vectype}; kwargs...) where {T, Tf, vectype}

This is the deflated version of the Newton Solver for F(x) = 0 with Jacobian J. It penalises the roots saved in defOp.roots. The other arguments are as for newton. See DeflationOperator for more informations.

Simplified calls are provided, for example when J is not passed. It then computed with finite differences.

Output:

solution:

history of residuals

flag of convergence

number of iterations

Newton for Fold / Hopf

PseudoArcLengthContinuation.newtonFold — Method

`newtonFold(F, J, foldpointguess::BorderedArray{vectype, T}, eigenvec, options::NewtonPar; Jt = nothing, d2F = nothing, normN = norm)`

This function turns an initial guess for a Fold point into a solution to the Fold problem based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is the parameter associated to the Fold point

dF = (x, p) -> d_xF(x, p) associated jacobian

foldpointguess initial guess (x0, p0) for the Fold point. It should be a BorderedArray as given by the function FoldPoint

eigenvec guess for the 0 eigenvector

options::NewtonPar

Optional arguments:

Jt = (x, p) -> transpose(d_xF(x, p)) associated jacobian transpose, it should be implemented in an efficient manner. For matrix-free methods, transpose is not readily available and the user must provide a dedicated method.

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) a bilinear operator representing the hessian of F. It has to provide an expression for d2F(x,p)[v1,v2].

normN = norm

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.newtonFold — Method

Simplified call to refine an initial guess for a Fold point. More precisely, the call is as follows

`newtonFold(F, J, br::ContResult, index::Int64, options::NewtonPar; Jt = nothing, d2F = nothing)`

where the optional argument Jt is the jacobian transpose and the Hessian is d2F. The parameters / options are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.newtonHopf — Method

`newtonHopf(F, J, hopfpointguess::BorderedArray{vectypeR, T}, eigenvec, eigenvec_ad, options::NewtonPar; Jt = nothing, d2F = nothing, normN = norm)`

This function turns an initial guess for a Hopf point into a solution to the Hopf problem based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is the parameter associated to the Hopf point

J = (x, p) -> d_xF(x, p) associated jacobian

hopfpointguess initial guess (x0, p0) for the Hopf point. It should a BorderedArray as given by the function HopfPoint.

eigenvec guess for the iω eigenvector

eigenvec_ad guess for the -iω eigenvector

options::NewtonPar

Optional arguments:

Jt = (x, p) -> transpose(d_xF(x, p)) associated jacobian transpose

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) a bilinear operator representing the hessian of F. It has to provide an expression for d2F(x,p)[v1, v2].

normN = norm

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.newtonHopf — Method

Simplified call to refine an initial guess for a Hopf point. More precisely, the call is as follows

`newtonHopf(F, J, br::ContResult, index::Int64, options; Jt = nothing, d2F = nothing, normN = norm)`

where the optional argument Jt is the jacobian transpose and the Hessian is d2F. The parameters are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Eigenvectors

This simplified call has been written when the eigenvectors are organised in a 2d Array evec where evec[:,2] is the second eigenvector in the list.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

Newton for Periodic Orbits

PseudoArcLengthContinuation.newton — Function

newton(prob::PeriodicOrbitTrapProblem, orbitguess, options::NewtonPar, linearPO = :BorderedLU; kwargs...)

This is the Newton Solver for computing a periodic orbit using a functional G based on Finite Differences and a Trapezoidal rule.

Arguments:

prob a problem of type PeriodicOrbitTrapProblem encoding the functional G

orbitguess a guess for the periodic orbit where orbitguess[end] is an estimate of the period of the orbit. It should be a vector of size N * M + 1 where M is the number of time slices, N is the dimension of the phase space. This must be compatible with the numbers N,M in prob.

options same as for the regular newton method

linearPO = :BorderedLU. Specify the choice of the linear algorithm, which must belong to [:FullLU, :BorderedLU, :FullMatrixFree, :BorderedMatrixFree]. This is used to select a way of inverting the jacobian dG of the functional G. For :FullLU, we use the default linear solver on a sparse matrix representation of dG. For :BorderedLU, we take advantage of the bordered shape of the linear solver and use LU decomposition to invert dG using a bordered linear solver. This is the default algorithm. For :FullMatrixFree, a matrix free linear solver is used for dG: note that a preconditioner is very likely required here because of the cyclic shape of dG which affects the convergence properties of GMRES. Finally, for :BorderedMatrixFree, a matrix free linear solver is used but for Jc only (see docs): it means that options.linsolver is used to invert Jc. These two Matrix-Free options thus expose different part of the jacobian dG in order to use specific preconditioners. For example, an ILU preconditioner on Jc could remove the constraints in dG and lead to poor convergence. Of course, for these last two methods, a preconditioner is likely to be required.

Output:

solution

history of residuals

flag of convergence

number of iterations

PseudoArcLengthContinuation.newton — Method

newton(prob::T, orbitguess, options::NewtonPar; kwargs...) where {T <: AbstractShootingProblem}

This is the Newton Solver for computing a periodic orbit using Shooting method.

Output:

solution

history of residuals

flag of convergence

number of iterations

Continuation

PseudoArcLengthContinuation.continuation — Function

continuation(F, J, x0, p0::Real, contParams::ContinuationPar; plot = false, normC = norm, dotPALC = (x,y) -> dot(x,y) / length(x), printSolution = norm, plotSolution = (x; kwargs...)->nothing, finaliseSolution = (z, tau, step, contResult) -> true, callbackN = (x, f, J, res, iteration, options; kwargs...) -> true, linearAlgo = BorderingBLS(), tangentAlgo = SecantPred(), verbosity = 0)

Compute the continuation curve associated to the functional F and its jacobian J.

Arguments:

F = (x, p) -> F(x, p) where p is the parameter for the continuation

J = (x, p) -> d_xF(x, p) its associated jacobian. It can be a matrix, a function or a callable struct.

u0 initial guess

p0 initial parameter, must be a real number

contParams parameters for continuatio. See ContinuationPar for more information about the options

plot = false whether to plot the solution while computing

printSolution = (x, p) -> norm(x) function used to plot in the continuation curve. It is also used in the way results are saved. It could be norm or x -> x[1]. This is also useful when saving several huge vectors is not possible for memory reasons (for example on GPU...).

plotSolution = (x; kwargs...) -> nothing function implementing the plot of the solution.

finaliseSolution = (z, tau, step, contResult) -> true Function called at the end of each continuation step. Can be used to alter the continuation procedure (stop it by returning false), saving personal data, plotting... The notations are $z=(x,p)$, tau is the tangent at z (see below), step is the index of the current continuation step and ContResult is the current branch. Note that you can have a better control over the continuation procedure by using an iterator, see Iterator Interface.

callbackN callback for newton iterations. see docs for newton. Can be used to change preconditioners

tangentAlgo = SecantPred() controls the algorithm use to predict the tangent along the curve of solutions or the corrector. Can be NaturalPred, SecantPred or BorderedPred.

linearAlgo = BorderingBLS(). Must belong to [MatrixBLS(), BorderingBLS(), MatrixFreeBLS()]. Used to control the way the extended linear system associated to the continuation problem is solved.

verbosity ∈ {0,1,2,3} controls the amount of information printed during the continuation process.

normC = norm norm used in the different Newton solves

dotPALC = (x,y) -> dot(x,y) / length(x), dot product used in the definition of the dot product (norm) $\|(u, p)\|^2_\theta$ in the constraint $N(x,p)$ (see below). This option can be used to remove the factor 1/length(x) for example in problems where the dimension of the state space changes (mesh adaptation, ...)

filename name of a file to save the computed branch during continuation. The identifier .jld2 will be appended to this filename

Outputs:

contres::ContResult composite type which contains the computed branch. See ContResult for more information.

u::BorderedArray the last solution computed on the branch

Controlling the parameter `linearAlgo`

In this simplified interface to continuation, the argument linearAlgo is internally overwritten to provide a valid argument to the algorithm. If you do not want this to happen, call directly continuation(F, J, x0, p0, contParams, linearAlgo; kwrags...).

Method

Bordered system of equations

The pseudo-arclength continuation method solves the equation $F(x, p) = 0$ (of dimension N) together with the pseudo-arclength constraint $N(x, p) = \frac{\theta}{length(x)} \langle x - x_0, \tau_0\rangle + (1 - \theta)\cdot(p - p_0)\cdot dp_0 - ds = 0$. In practice, the curve $\gamma$ is parametrised by $s$ so that $\gamma(s) = (x(s), p(s))$ is a curve of solutions to $F(x, p)$. This formulation allows to pass turning points (where the implicit theorem fails). In the previous formula, $(x_0, p_0)$ is a solution for a given $s_0$, $(\tau_0, dp_0)$ is the tangent to the curve at $s_0$. Hence, to compute the curve of solutions, we need solve an equation of dimension N+1 which is called a Bordered system.

Parameter `theta`

The parameter theta in the struct ContinuationParis very important. It should be tuned for the continuation to work properly especially in the case of large problems where the $\langle x - x_0, \tau_0\rangle$ component in the constraint might be favoured too much.

The parameter ds is adjusted internally depending on the number of Newton iterations and other factors. See the function stepSizeControl for more information. An important parameter to adjust the magnitude of this adaptation is the parameter a in the struct ContinuationPar.

Algorithm

The algorithm works as follows:

Start from a known solution $(x_0, p_0,\tau_0 ,dp_0)$

Predictor set $(x_1, p_1) = (x_0, p_0) + ds\cdot (\tau_0, dp_0)$

Corrector solve $F(x, p)=0,\ N(x, p)=0$ with a (Bordered) Newton Solver with guess $(x_1, p_1)$.

if Newton in 3. did not converge, put ds/2 ⟶ ds and go to 1.

New tangent Compute $(\tau_1, dp_1)$, set $(x_0, p_0, \tau_0, dp_0) = (x_1, p_1, \tau_1, dp_1)$ and return to step 2

Natural continuation

We speak of natural continuation when we do not consider the constraint $N(x, p)=0$. Knowing $(x_0, p_0)$, we use $x_0$ as a guess for solving $F(x, p_1)=0$ with $p_1$ close to $p_0$. Again, this fails at Turning points but it can be faster to compute than the constrained case. This is set by the option tangentAlgo = NaturalPred() in continuation.

Tangent computation (step 4)

There are various ways to compute $(\tau_1, p_1)$. The first one is called secant and is parametrised by the option tangentAlgo = SecantPred(). It is computed by $(\tau_1, p_1) = (z_1, p_1) - (z_0, p_0)$ and normalised by the norm $\|(u, p)\|^2_\theta = \frac{\theta}{length(u)} \langle u,u\rangle + (1 - \theta)\cdot p^2$. Another method is to compute $(\tau_1, p_1)$ by solving solving the bordered linear system $\begin{bmatrix} F_x & F_p ; \ \frac{\theta}{length(x)}\tau_0 & (1-\theta)p_0\end{bmatrix}\begin{bmatrix}\tau_1 ; p_1\end{bmatrix} =\begin{bmatrix}0 ; 1\end{bmatrix}$ ; it is set by the option tangentAlgo = BorderedPred().

Bordered linear solver

When solving the Bordered system $F(x, p) = 0,\ N(x, p)=0$, one faces the issue of solving the Bordered linear system $\begin{bmatrix} J & a ; b^T & c\end{bmatrix}\begin{bmatrix}X ; y\end{bmatrix} =\begin{bmatrix}R ; n\end{bmatrix}$. This can be solved in many ways via bordering (which requires two Jacobian inverses), by forming the bordered matrix (which works well for sparse matrices) or by using a full Matrix Free formulation. The choice of method is set by the argument linearAlgo. Have a look at the struct linearBorderedSolver for more information.

Linear Algebra

Let us discuss here more about the norm and dot product. First, the option normC gives a norm that is used to evaluate the residual in the following way: $max(normC(F(x,p)), \|N(x,p)\|)<tol$. It is thus used as a stopping criterion for a Newton algorithm. The dot product (resp. norm) used in $N$ and in the (iterative) linear solvers is LinearAlgebra.dot (resp. LinearAlgebra.norm). It can be changed by importing these functions and redefining it. Not that by default, the $L^2$ norm is used. These details are important because of the constraint $N$ which incorporates the factor length. For some custom composite type implementing a Vector space, the dot product could already incorporates the length factor in which case you should either redefine the dot product or change $\theta$.

Step size control

As explained above, each time the corrector phased failed, the step size $ds$ is halved. This has the disavantage of having lost Newton iterations (which costs time) and impose small steps (which can be slow as well). To prevent this, the step size is controlled internally with the idea of having a constant number of Newton iterations per point. This is in part controlled by the aggressiveness factor a in ContinuationPar. Further tuning is performed by using doArcLengthScaling=true in ContinuationPar. This adjusts internally $\theta$ so that the relative contributions of $x$ and $p$ are balanced in the constraint $N$.

Continuation for Fold / Hopf

PseudoArcLengthContinuation.continuationFold — Method

Codim 2 continuation of Fold points. This function turns an initial guess for a Fold point into a curve of Fold points based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p1, p2) -> F(x, p1, p2) where p is the parameter associated to the Fold point

J = (x, p1, p2) -> d_xF(x, p1, p2) associated jacobian

foldpointguess initial guess (x0, p10) for the Fold point. It should be a BorderedArray as given by the function FoldPoint

p2 parameter p2 for which foldpointguess is a good guess

eigenvec guess for the 0 eigenvector at p1_0

options::NewtonPar

Optional arguments:

Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) associated jacobian transpose

d2F = p2 -> ((x, p1, v1, v2) -> d2F(x, p1, p2, v1, v2)) this is the hessian of F computed at (x, p1, p2) and evaluated at (v1, v2).

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.continuationFold — Method

Simplified call for continuation of Fold point. More precisely, the call is as follows continuationFold(F, J, br::ContResult, index::Int64, options; Jt = nothing, d2F = p2 -> nothing) where the parameters are as for continuationFold except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.continuationHopf — Method

Simplified call for continuation of Hopf point. More precisely, the call is as follows continuationHopf(F, J, br::ContResult, index::Int64, options; Jt = nothing, d2F = p2 -> nothing) where the parameters are as for continuationHopf except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Simplified calls are also provided but at the cost of using finite differences.

Eigenvectors

This simplified call has been written when the eigenvectors are organised in a 2d Array evec where evec[:,2] is the second eigenvector in the list.

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

PseudoArcLengthContinuation.continuationHopf — Method

codim 2 continuation of Hopf points. This function turns an initial guess for a Hopf point into a curve of Hopf points based on a Minimally Augmented formulation. The arguments are as follows

(x, p1, p2)-> F(x, p1, p2) where p is the parameter associated to the hopf point

J = (x, p1, p2)-> d_xF(x, p1, p2) associated jacobian

hopfpointguess initial guess (x0, p10) for the Hopf point. It should be a Vector or a BorderedArray

p2 parameter p2 for which hopfpointguess is a good guess

eigenvec guess for the iω eigenvector at p1_0

eigenvec_ad guess for the -iω eigenvector at p1_0

options::NewtonPar

Optional arguments:

Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) associated jacobian transpose

d2F = p2 -> ((x, p1, v1, v2) -> d2F(x, p1, p2, v1, v2)) this is the hessian of F computed at (x, p1, p2) and evaluated at (v1, v2).

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

Jacobian tranpose

The adjoint of the jacobian J is computed internally when Jt = nothing by using tranpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) otherwise the jacobian will be computed twice!

Hessian

The hessian of F, when d2F is not passed, is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

Continuation for periodic orbits

PseudoArcLengthContinuation.continuationPOTrap — Function

continuationPOTrap(prob, orbitguess, p0::Real, contParams::ContinuationPar, linearPO = :BorderedLU; printSolution = (u,p) -> u[end], kwargs...)

This is the continuation routine for computing a periodic orbit using a functional G based on Finite Differences and a Trapezoidal rule.

Arguments

p -> prob(p) is a family such that prob(p)::PeriodicOrbitTrapProblem encodes the functional G

orbitguess a guess for the periodic orbit where orbitguess[end] is an estimate of the period of the orbit. It could be a vector of size N * M + 1 where M is the number of time slices, N is the dimension of the phase space. This must be compatible with the numbers N, M in prob.

p0 initial parameter, must be a real number

contParams same as for the regular continuation method

linearPO = :BorderedLU. Same as newton when applied to PeriodicOrbitTrapProblem.

Note that by default, the methods prints the period of the periodic orbit as function of the parameter. This can be changed by providing your printSolution argument.

PseudoArcLengthContinuation.continuationPOShooting — Method

continuationPOShooting(prob, orbitguess, p0::Real, contParams::ContinuationPar; printPeriod = true, kwargs...)

This is the continuation routine for computing a periodic orbit using a functional G based on a Shooting method.

Arguments

p -> prob(p) is a function or family such that prob(p)::AbstractShootingProblem encodes the functional G

orbitguess a guess for the periodic orbit. For the type of orbitguess, please see the information concerning ShootingProblem and PoincareShootingProblem.

p0 initial parameter, must be a real number

contParams same as for the regular continuation method

printPeriod in the case of Poincaré Shooting, plot the period of the cycle.

Plotting

PseudoArcLengthContinuation.plotBranch — Method

plotBranch(contres::ContResult, plot_fold = true; kwargs...)

Plot the branch of solutions from a ContResult. You can also pass parameters like plotBranch(br, marker = :dot). For the continuation diagram, the legend is as follows (:fold => :black, :hopf => :red, :bp => :blue, :nd => :magenta, :none => :yellow)

PseudoArcLengthContinuation.plotBranch! — Method

plotBranch!(contres, plot_fold = true; putbifptlegend = true, filterbifpoints::Bool = false, kwargs...)

Append to the current plot, the plot of the branch of solutions from a ContResult. You can also pass parameters like plotBranch!(br, marker = :dot). Options to filter the bifurcation points (which are mostly guesses) are provided. For example, filterbifpoints = true merges the nearby Fold and Branch points.

PseudoArcLengthContinuation.plotBranch — Method

plotBranch(contres::ContResult, plot_fold = true; kwargs...)

Plot the branch of solutions from a ContResult. You can also pass parameters like plotBranch(br, marker = :dot). For the continuation diagram, the legend is as follows (:fold => :black, :hopf => :red, :bp => :blue, :nd => :magenta, :none => :yellow)

plotBranch(brs::Vector, plot_fold = true; putbifptlegend = true, filterbifpoints = false, kwargs...)

Plot all the branches contained in brs in a single figure. Convenient when many bifurcation diagram have been computed.