Library

Library

Structs

PseudoArcLengthContinuation.PeriodicOrbitTrap — Type.

pb = PeriodicOrbitTrap(F, J, ϕ, xπ, M::Int, linsolve)

This structure implements Finite Differences based on Trapezoidal rule to locate periodic orbits. The arguements are as follows

F vector field

J jacobian of the vector field

ϕ used for the Poincare section

xπ used for the Poincare section

M::Int number of slices in [0,2π]

linsolve <: LinearSolver linear solver

You can then call pb(orbitguess) to apply the functional to a guess. Note that orbitguess must be of size M * N + 1 where N is the number of unknowns in the state space and orbitguess[M*N+1] is an estimate of the period of the limit cycle.

The scheme is as follows, one look for T = x[end] and $x_{i+1} - x_{i} - \frac{h}{2} \left(F(x_{i+1}) + F(x_i)\right) = 0$

where h = T/M. Finally, the phase of the periodic orbit is constraint by

$\langle x[1] - x\pi, \phi\rangle.$

source

Newton

PseudoArcLengthContinuation.newton — Function.

newton(F, J, x0, options, normN = norm)

This is the Newton Solver for F(x) = 0 with Jacobian J and initial guess x0. The function normN allows to specify a norm for the convergence criteria. It is important to set the linear solver options.linsolve properly depending on your problem. This solver is used to solve $J(x)u = -F(x)$ in the Newton step. You can for example use linsolve = Default() which is the operator backslash: it works well for Sparse / Dense matrices. Iterative solver (GMRES) are also provided. You should implement your own solver for maximal efficiency. This is quite easy to do, have a look at src/LinearSolver.jl. The functions or callable which need to be passed are as follows:

x -> F(x) functional whose zeros are looked for. In particular, it is not inplace,

dF(x) = x -> J(x) compute the jacobian of F at x. It is then passed to options.linsolve. The Jacobian can be a matrix or an out of place function.

Simplified calls are provided, for example when J is not passed. It then computed with finite differences.

Output:

solution:

history of residuals

flag of convergence

number of iterations

source

PseudoArcLengthContinuation.newtonDeflated — Function.

newtonDeflated(Fhandle::Function, Jhandle, x0, options:: NewtonPar{T}, defOp::DeflationOperator{T, vectype})

This is the deflated version of the Newton Solver. It penalises the roots saved in defOp.roots

source

Newton for Fold / Hopf

PseudoArcLengthContinuation.newtonFold — Method.

This function turns an initial guess for a Fold point into a solution to the Fold problem based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p) -> F(x, p) where p is the parameter associated to the Fold point

dF = (x, p) -> d_xF(x, p) associated jacobian

dFt = (x, p) -> transpose(d_xF(x, p)) associated jacobian, it should be implemented in an efficient manner. For matrix-free methods, tranpose is not readily available.

d2F = (x, p, v1, v2) -> d2F(x, p, v1, v2) a bilinear operator representing the hessian of F. It has to provide an expression for d2F(x,p)[v1,v2].

foldpointguess initial guess (x0, p0) for the Fold point. It should be a BorderedArray as given by the function FoldPoint

eigenvec guess for the 0 eigenvector

options::NewtonPar

source

PseudoArcLengthContinuation.newtonFold — Method.

Simplified call to refine an initial guess for a Fold point. More precisely, the call is as follows

`newtonFold(F, J, Jt, br::ContResult, index::Int64, options)`

or

`newtonFold(F, J, Jt, d2F, br::ContResult, index::Int64, options)`

when the Hessian d2F is known. The parameters are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

source

Continuation

PseudoArcLengthContinuation.continuation — Function.

continuation(F, J, u0, p0::Real, contParams::ContinuationPar; plot = false, normC = norm, printsolution = norm, plotsolution::Function = (x; kwargs...)->nothing, finaliseSolution::Function = (z, tau, step, contResult) -> true, linearalgo = :bordering, verbosity = 2)

Compute the continuation curve associated to the functional F and its jacobian J. The parameters are as follows

F = (x, p) -> F(x, p) where p is the parameter for the continuation

J = (x, p) -> d_xF(x, p) its associated jacobian. It can be a matrix, a function or a callable struct.

u0 initial guess

contParams parameters for continuation, with struct ContinuationPar

plot = false whether to plot the solution while computing

printsolution::Function = norm function used to plot in the continuation curve, e.g. norm or x -> x[1]

plotsolution::Function = (x; kwargs...) -> nothing function implementing the plot of the solution.

finaliseSolution::Function = (z, tau, step, contResult) -> true Function called at the end of each continuation step. Can be used to alter the continuation step (stop it by returning false) or saving personal data...

linearalgo = :bordering. Must belong to [:bordering, :full]. Used to control the way the extended linear system associated to the continuation problem is solved.

verbosity controls the amount of information printed during the continuation process.

normC = norm norm used in the different Newton solves

The function outputs

contres::ContResult structure which contains the computed branch

u::BorderedArray the last solution computed on the branch

Method

Bordered system of equations

The pseudo arclength continuation method solves the equation $F(x,p) = 0$ (or dimension N) together with the pseudo-arclength constraint $N(x, p) = \frac{\theta}{length(x)} \langle x - x_0, \tau_0\rangle + (1 - \theta)\cdot(p - p_0)\cdot dp_0 - ds = 0$. In practice, the curve is parametrised by $s$ so that $(x(s),p(s))$ is a curve of solutions to $F(x,p)$. This formulation allows to pass turning points (where the implicit theorem fails). In the previous formula, $(x_0, p_0)$ is a solution for a given $s_0$, $(\tau_0, dp_0)$ is the tangent to the curve at $s_0$. Hence, to compute the curve of solutions, we need solve an equation of dimension N+1 which is called a Bordered system.

Parameter `theta`

The parameter theta in the struct ContinuationParis very important. It should be tuned for the continuation to work properly especially in the case of large problems where the $\langle x - x_0, \tau_0\rangle$ component in the constraint might be favoured too much.

The parameter ds is adjusted internally depending on the number of Newton iterations and other factors. See the function stepSizeControl for more information. An important parameter to adjust the magnitude of this adaptation is the parameter a in the struct ContinuationPar.

Algorithm

The algorithm works as follows:

Start from a known solution $(x_0,p_0,\tau_0,dp_0)$

Predictor set $(x_1,p_1) = (x_0,p_0) + ds\cdot (\tau_0,dp_0)$

Corrector solve $F(x,p)=0,\ N(x,p)=0$ with a (Bordered) Newton Solver.

New tangent Compute $(\tau_1,dp_1)$, set $(x_0,p_0,\tau_0,dp_0)=(x_1,p_1,\tau_1,dp_1)$ and return to step 2

Natural continuation

We speak of natural continuation when we do not consider the constraint $N(x,p)=0$. Knowing $(x_0,p_0)$, we use $x_0$ as a guess for solving $F(x,p_1)=0$ with $p_1$ close to $p_0$. Again, this will fail at Turning points but it can be faster to compute than the constrained case. This is set by the field natural in the struct ContinuationPar

Tangent computation (step 4)

There are various ways to compute $(\tau_1,p_1)$. The first one is called secant and is parametrised by the field secant in the struct ContinuationPar. It is computed by $(\tau_1,p_1) = (z_1,p_1) - (z_0,p_0)$ and normalised by the norm $\|u,p\|^2_\theta = \frac{\theta}{length(u)} \langle u,u\rangle + (1 - \theta)\cdot p^2$. If secant is set to false, another method is use computing $(\tau_1,p_1)$ by solving a bordered linear system, see the function getTangentBordered for more information.

Bordered linear solver

When solving the Bordered system $F(x,p) = 0,\ N(x, p)=0$, one faces the issue of solving the Bordered linear system $\begin{bmatrix} J & a ; b^T & c\end{bmatrix}\begin{bmatrix}X ; y\end{bmatrix} =\begin{bmatrix}R ; n\end{bmatrix}$. This can be solved in many ways via bordering (which requires two Jacobian inverses) or by forming the bordered matrix (which works well for sparse matrices). The choice of method is set by the argument linearalgo. Have a look at the function linearBorderedSolver for more information.

source

Continuation for Fold / Hopf

PseudoArcLengthContinuation.continuationFold — Method.

codim 2 continuation of Fold points. This function turns an initial guess for a Fold point into a curve of Fold points based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p1, p2) -> F(x, p1, p2) where p is the parameter associated to the Fold point

J = (x, p1, p2) -> d_xF(x, p1, p2) associated jacobian

Jt = (x, p1, p2) -> transpose(d_xF(x, p1, p2)) associated jacobian

d2F = (x, p1, p2, v1, v2) -> d2F(x, p1, p2, v1, v2) this is the hessian of F computed at (x, p1, p2) and evaluated at (v1, v2).

foldpointguess initial guess (x0, p10) for the Fold point. It should be a BorderedArray as given by the function FoldPoint

p2 parameter p2 for which foldpointguess is a good guess

eigenvec guess for the 0 eigenvector at p1_0

options::NewtonPar

source

PseudoArcLengthContinuation.continuationFold — Method.

codim 2 continuation of Fold points. This function turns an initial guess for a Fold point into a curve of Fold points based on a Minimally Augmented formulation. The arguments are as follows

F = (x, p1, p2) -> F(x, p1, p2) where p is the parameter associated to the Fold point

J = (x, p1, p2) -> d_xF(x, p1, p2) associated jacobian

foldpointguess initial guess (x0, p10) for the Fold point. It should be a BorderedArray as given by the function FoldPoint

p2 parameter p2 for which foldpointguess is a good guess

eigenvec guess for the 0 eigenvector at p1_0

options::NewtonPar

Hessian

The hessian of `F` in this case is computed with Finite differences. This can be slow for many variables, e.g. ~1e6

source

PseudoArcLengthContinuation.continuationFold — Method.

Simplified call for continuation of Fold point. More precisely, the call is as follows continuationFold(F, J, Jt, d2F, br::ContResult, index::Int64, options) where the parameters are as for continuationFold except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine.

Simplified calls are also provided but at the cost of using finite differences.

source

Plotting

PseudoArcLengthContinuation.plotBranch — Method.

Plot the branch of solutions from a ContResult. You can also pass parameters like plotBranch(br, marker = :dot). For the continuation diagram, the legend is as follows (:fold => :black, :hopf => :red, :bp => :blue, :nd => :magenta, :none => :yellow)

source

PseudoArcLengthContinuation.plotBranch! — Method.

Append to the current plot the plot of the branch of solutions from a ContResult. You can also pass parameters like plotBranch!(br, marker = :dot)

source

PseudoArcLengthContinuation.plotBranch — Method.

Plot all the branches contained in brs in a single figure. Convenient when many bifurcation diagram have been computed.

source